后台提交订单完成

+++ res/pieces/pieces-biz/src/main/java/com/pieces/biz/controller/OrderController.java
@@ -154,7 +154,7 @@ public class OrderController extends BaseController {
 			oc.setOriginOf(ec.getOrigin());
 			oc.setExpectDate(ec.getExpectDate());
 			oc.setAmount(ec.getAmount());
-			oc.setPrice(ec.getMyPrice().floatValue());
+			oc.setPrice(ec.getMyPrice());
 			oc.setSubtotal(oc.getAmount()*oc.getPrice());
 			oc.setEnquiryCommodityId(ec.getId());
 			oc.setOrderId(null);
